refactor: 공통 인터페이스 처리 및 파일 네이밍 개선
KB3-84
