--- src/types/common.ts
@@ -1,2 +1,0 @@
-export type PetType = '강아지' | '고양이' | '햄스터' | '조류' | '어류' | '파충류';
-export type PetGender = '남아' | '여아' | '중성';
