+++ src/types/form.ts
@@ -1,3 +1,13 @@
+export type PetType = '강아지' | '고양이' | '햄스터' | '조류' | '어류' | '파충류';
+export type PetGender = '남아' | '여아' | '중성';
+
+export interface PetForm {
+  name: string;
+  species: PetType;
+  gender: PetGender;
+  birthDate: string;
+}
+
 export interface BaseFieldProps {
   label?: string;
 }
